pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "A tool to convert markdown files to professionally styled PDFs"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Nexus Team" }
]
keywords = ["markdown", "pdf", "converter", "documentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "click>=8.0.0",
    "markdown>=3.3.0",
    "jinja2>=3.0.0",
    "pdfkit>=1.0.0",
    "PyYAML>=6.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]

[project.scripts]
nexus = "nexus.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
